# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "javascript_seminar_meeting",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:4200/bbbbridge/api/v1/bbb"
    }
  ],
  "paths": {
...
    "/monitoring/get_meetings": {
      "get": {
        "summary": "Gets meetings",
        "tags": [
          "monitoring"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/monitoring/get_meeting_info": {
      "post": {
        "summary": "Gets meeting Info",
        "tags": [
          "monitoring"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "meetingId"
                ],
                "properties": {
                  "meetingId": {
                    "type": "string",
                    "description": "The meeting ID that identifies the meeting you are attempting to check on."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/monitoring/is_meeting_running": {
      "post": {
        "summary": "Checks if meeting is running",
        "tags": [
          "monitoring"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "meetingId"
                ],
                "properties": {
                  "meetingId": {
                    "type": "string",
                    "description": "The meeting ID that identifies the meeting you are attempting to check on."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}